reduce maxPlacementScale
This makes placement 2-4x faster.

Sets the maxPlacementScale to match the intended zoom range for the
tiles. For most tiles this is `1`. For the highest zoom level tile this
is the difference between that zoom and the map's max zoom.

This makes sense now that tiles are loaded at every zoom level, and up
to a higher zoom level.