only interpolate for tile's ideal zoom range
This limits the depth of the interpolation for line symbols based on
what the tile's ideal zoom range is. For most tiles this will limit the
interpolation to a depth of 1.

This improves performance considerably by exponentially reducing the
number of symbols being placed for that layer (which also reduces the
number of symbols layer symbols will be checked against).

It also improves label density because less deep labels are blocking
ones that could show up earlier.