make collision detection work in perspective view
Labels should mostly not overlap in perspective view.

The problem is that the angle from the camera to the surface is
different at the lower edge compared to the upper edge. This means that
labels near the upper edge cover a larger area than at the bottom edge,
even after they have been scaled for size. The only way to really fix
this is to redo placement when panning in the y direction.

The current implementation seems like it might be ok, so let's try this
for a while and see how it works.