switch to unpacked representation of the tile id
The standard representation of a tile id used to be a number with
x, y, z, and w packed into it. It was being constantly unpacked
everywhere it was being used.

This switches to mainly using an object to hold the id.
The packed representation is only used as a key to index the tile.