Replace temp tokens for Mapbox tiles with latest token
For raster and vector tiles, the URLs defined in the TileJSON include the
access token the json was loaded with. For temporary tokens, tiles could
be requested with a stale token that had since been refreshed, but pinned
by the TileJSON.

Now, we replace a temp token in a tile url for mapbox sources with the
most current access token. Non-mapbox sources and non-temp token are
ignored.