Stub out Map#_rerender in paint benchmarks
Fixes #5543. Internal rendering changes were causing one of the conditions on the call to _rerender inside _render to be true, whereas before it was apparently always false. _rerender caused an animation frame to be requested, but then the resulting frameId would be reset via the benchmark's own call to _render. The result was a backlog of animation frame requests, some of which completed after teardown.