skip symbol fading for tiles immediately after reload
The new symbol fading was causing flickering for symbols animated with
`setData`. When the new location of the symbol was far enough away from
the old it would not identify them as being the same icon, resulting in
fading. Solving this properly would include figuring out a good way to
fix this matching problem.

Instead, this commit works around the problem by skipping fading for any
tiles that have just been reloaded.